+++ CHANGELOG.md
@@ -12,10 +12,12 @@ For a full diff see [`dcd4cfb...main`][dcd4cfb...main].
 
 * Added `Json` ([#2]), by [@localheinz]
 * Added `Schema` ([#3]), by [@localheinz]
+* Added `Decoder` ([#5]), by [@localheinz]
 
 [dcd4cfb...main]: https://github.com/ergebnis/json-schema-validator/compare/dcd4cfb...main
 
 [#2]: https://github.com/ergebnis/json-schema-validator/pull/2
 [#3]: https://github.com/ergebnis/json-schema-validator/pull/3
+[#5]: https://github.com/ergebnis/json-schema-validator/pull/5
 
 [@localheinz]: https://github.com/localheinz